CSGO ShotDEX Client
"CSGO ShotDEX" is a comprehensive tool designed to enhance your gaming experience within Counter-Strike: Global Offensive (CS:GO). With an array of features tailored for precision, strategy, and customization, this tool provides seamless integration and ease of use for gamers seeking additional functionalities within the game.
Features:

    -Aimbot: Customize mouse positioning, delay, and hitbox selection for precision targeting.
    -Recoil Control System (RCS): Manage recoil with custom pitch and yaw settings.
    -Triggerbot: Automatically shoot at targets in your crosshair.
    -Backtrack: Manipulate tickrate for historical record storage.
    -Fake Latency: Introduce fake latency to enhance backtrack capabilities.
    -Bunnyhop: Enable automatic strafing for enhanced movement.
    -Thirdperson: Activate a third-person view with customizable settings.
    -Crosshairs: Display various crosshair styles for improved targeting.
    ...and a comprehensive list of additional features for an enriched gaming experience.

Ease of Use:

"CSGO ShotDEX" is designed for simplicity. Just start the .exe file, and you're ready to experience enhanced gameplay.
Extras:

   - Threadsafe Streamproof Render Wrapper: Deprecated surface rendering.
   - Icon Extraction: Extract game icons to images.
   - Game Callbacks and Logging Console: Improve game communication and logging.
   - Entity Cache and Game Events: Streamline entity data and game event handling.
   - Runtime Shutdown and Streamproof Rendering: Optimize performance and security during gameplay.

This tool is crafted to offer an extensive range of features while ensuring ease of use for gamers of all levels.